vendor: Add Ancient font package
Fonts taken from dafont.com and free to use for personal use and not
commercial use.

https://www.dafont.com/

According to 18 USCS § 31, term "used for commercial purposes" means the
carriage of persons or property for any fare, fee, rate, charge or other
consideration, or directly or indirectly in connection with any business,
or other undertaking intended for profit.

https://definitions.uslegal.com/u/used-for-commercial-purposes/

Since DU is not and will never be a for profit, we're in the clear :-)

- Fonts xml files generated for these font zips thanks to @USA-RedDragon

Change-Id: I1440d6c51ee25c7e465e83b2f4ec5099f03e741d